feat(bt): raise BR/EDR tx power to +9dBm, fix Bf report, slim SPP callback
- set esp_bredr_tx_power_set(P9,P9) after SerialBT.begin (default cap was +3dBm)
- report real 2048-byte InputBuffer availability in Bf for CLIENT_BT
  (was 512-based and could go negative, breaking sender flow control)
- defer BT connect/disconnect messages and _btclient update from SPP
  callback (BT stack task, core 0) to BTConfig::handle() on core 1;
  _btclient String -> char[18], removing cross-core heap race and
  potential SPP write deadlock in callback
- document BT signal/discovery findings in BLUETOOTH_IMPROVEMENT.md

+++ Grbl_Esp32/src/Serial.h
@@ -48,7 +48,7 @@ void client_init();
 void client_reset_read_buffer(uint8_t client);
 
 // Returns the number of bytes available in the RX serial buffer.
-uint8_t client_get_rx_buffer_available(uint8_t client);
+int client_get_rx_buffer_available(uint8_t client);
 
 void execute_realtime_command(Cmd command, uint8_t client);
 bool is_realtime_command(uint8_t data);
